+++ src/core/client/admin/routes/Configure/sections/Moderation/ModerationConfigContainer.tsx
@@ -15,7 +15,7 @@ import { ModerationConfigContainer_settings as SettingsData } from "coral-admin/
 import AkismetConfig from "./AkismetConfig";
 import NewCommentersConfig from "./NewCommentersConfig";
 import PerspectiveConfig from "./PerspectiveConfig";
-import PreModerationConfig from "./PreModerationConfig";
+import PreModerationConfig from "./PreModerationConfigContainer";
 import RecentCommentHistoryConfig from "./RecentCommentHistoryConfig";
 
 interface Props {
@@ -55,8 +55,8 @@ const enhanced = withFragmentContainer<Props>({
     fragment ModerationConfigContainer_settings on Settings {
       ...AkismetConfig_formValues @relay(mask: false)
       ...PerspectiveConfig_formValues @relay(mask: false)
-      ...PreModerationConfig_formValues @relay(mask: false)
-      ...PreModerateAllCommentsConfigContainer_settings
+      ...PreModerationConfigContainer_formValues @relay(mask: false)
+      ...PreModerationConfigContainer_settings
       ...RecentCommentHistoryConfig_formValues @relay(mask: false)
       ...NewCommentersConfigContainer_settings @relay(mask: false)
     }
